@@ -77,118 +79,115 @@
     "canvas-confetti": "^1.9.3",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.13",
-    "diff": "^7.0.0",
-    "framer-motion": "^11.18.0",
-    "globby": "^14.0.2",
+    "diff": "^8.0.2",
+    "framer-motion": "^12.15.0",
+    "globby": "^14.1.0",
     "gray-matter": "^4.0.3",
     "html-to-text": "^9.0.5",
-    "ioredis": "^5.4.2",
-    "isomorphic-dompurify": "^2.20.0",
+    "ioredis": "^5.6.1",
+    "isomorphic-dompurify": "^2.25.0",
     "js-cookie": "^3.0.5",
-    "jsdom": "^26.0.0",
+    "jsdom": "^26.1.0",
     "jsonwebtoken": "^9.0.2",
     "localforage": "^1.10.0",
-    "lucide-react": "^0.469.0",
+    "lucide-react": "^0.511.0",
     "lz-string": "^1.5.0",
-    "next": "15.1.6",
+    "next": "15.3.3",
     "next-mdx-remote": "^5.0.0",
-    "next-nprogress-bar": "^2.4.3",
-    "next-themes": "^0.4.4",
-    "node-cron": "^3.0.3",
-    "nodemailer": "^6.9.16",
+    "next-nprogress-bar": "^2.4.7",
+    "next-themes": "^0.4.6",
+    "node-cron": "^4.1.0",
+    "nodemailer": "^7.0.3",
     "nodemailer-smtp-transport": "^2.7.4",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "react-dropzone": "^14.3.5",
-    "react-hook-form": "^7.54.2",
-    "react-hot-toast": "^2.5.1",
-    "react-hotkeys-hook": "^4.6.1",
-    "react-image-crop": "^11.0.7",
-    "refractor": "^4.8.1",
-    "rehype-prism-plus": "^2.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-dropzone": "^14.3.8",
+    "react-hook-form": "^7.57.0",
+    "react-hot-toast": "^2.5.2",
+    "react-hotkeys-hook": "^5.1.0",
+    "react-image-crop": "^11.0.10",
+    "refractor": "^5.0.0",
+    "rehype-prism-plus": "^2.0.1",
     "rehype-sanitize": "^6.0.0",
     "rehype-stringify": "^10.0.1",
     "remark-frontmatter": "^5.0.0",
-    "remark-gfm": "^4.0.0",
+    "remark-gfm": "^4.0.1",
     "remark-parse": "^11.0.0",
-    "remark-rehype": "^11.1.1",
+    "remark-rehype": "^11.1.2",
     "rimraf": "^6.0.1",
-    "sharp": "^0.33.5",
-    "sugar-high": "^0.7.5",
-    "tailwind-merge": "^2.6.0",
+    "sharp": "^0.34.2",
+    "sugar-high": "^0.9.3",
+    "tailwind-merge": "^3.3.0",
     "typed.js": "^2.1.0",
     "unified": "^11.0.5",
     "unist-util-visit": "^5.0.0",
     "use-debounce": "^10.0.4",
-    "yet-another-react-lightbox": "^3.21.7",
-    "zod": "^3.24.1",
-    "zod-form-data": "^2.0.5",
-    "zustand": "^5.0.3"
+    "yet-another-react-lightbox": "^3.23.2",
+    "zod": "^3.25.48",
+    "zod-form-data": "^3.0.0",
+    "zustand": "^5.0.5"
   },
   "devDependencies": {
-    "@milkdown/core": "7.5.9",
-    "@milkdown/crepe": "^7.5.9",
-    "@milkdown/ctx": "7.5.9",
-    "@milkdown/exception": "7.5.9",
-    "@milkdown/kit": "^7.5.9",
-    "@milkdown/plugin-block": "7.5.9",
-    "@milkdown/plugin-clipboard": "7.5.9",
-    "@milkdown/plugin-cursor": "7.5.9",
-    "@milkdown/plugin-diagram": "7.5.9",
-    "@milkdown/plugin-emoji": "7.5.9",
-    "@milkdown/plugin-history": "7.5.9",
-    "@milkdown/plugin-indent": "7.5.9",
-    "@milkdown/plugin-listener": "7.5.9",
-    "@milkdown/plugin-math": "7.5.9",
-    "@milkdown/plugin-prism": "7.5.9",
-    "@milkdown/plugin-slash": "7.5.9",
-    "@milkdown/plugin-tooltip": "7.5.9",
-    "@milkdown/plugin-trailing": "7.5.9",
-    "@milkdown/plugin-upload": "7.5.9",
-    "@milkdown/preset-commonmark": "7.5.9",
-    "@milkdown/preset-gfm": "7.5.9",
-    "@milkdown/prose": "7.5.9",
-    "@milkdown/react": "7.5.9",
-    "@milkdown/theme-nord": "7.5.9",
-    "@milkdown/transformer": "7.5.9",
-    "@milkdown/utils": "7.5.9",
-    "@prosemirror-adapter/react": "^0.4.0",
+    "@milkdown/core": "7.12.1",
+    "@milkdown/crepe": "^7.12.1",
+    "@milkdown/ctx": "7.12.1",
+    "@milkdown/exception": "7.12.1",
+    "@milkdown/kit": "^7.12.1",
+    "@milkdown/plugin-block": "7.12.1",
+    "@milkdown/plugin-clipboard": "7.12.1",
+    "@milkdown/plugin-cursor": "7.12.1",
+    "@milkdown/plugin-emoji": "7.12.1",
+    "@milkdown/plugin-history": "7.12.1",
+    "@milkdown/plugin-indent": "7.12.1",
+    "@milkdown/plugin-listener": "7.12.1",
+    "@milkdown/plugin-prism": "7.12.1",
+    "@milkdown/plugin-slash": "7.12.1",
+    "@milkdown/plugin-tooltip": "7.12.1",
+    "@milkdown/plugin-trailing": "7.12.1",
+    "@milkdown/plugin-upload": "7.12.1",
+    "@milkdown/preset-commonmark": "7.12.1",
+    "@milkdown/preset-gfm": "7.12.1",
+    "@milkdown/prose": "7.12.1",
+    "@milkdown/react": "7.12.1",
+    "@milkdown/theme-nord": "7.12.1",
+    "@milkdown/transformer": "7.12.1",
+    "@milkdown/utils": "7.12.1",
+    "@prosemirror-adapter/react": "^0.4.1",
     "@tailwindcss/typography": "^0.5.16",
     "@types/canvas-confetti": "^1.9.0",
-    "@types/diff": "^6.0.0",
     "@types/html-to-text": "^9.0.4",
     "@types/js-cookie": "^3.0.6",
-    "@types/jsonwebtoken": "^9.0.7",
-    "@types/node": "22.10.2",
-    "@types/react": "19.0.2",
-    "@types/react-dom": "19.0.2",
-    "autoprefixer": "^10.4.20",
-    "dotenv": "^16.4.7",
-    "eslint": "^9.18.0",
-    "eslint-config-next": "^15.1.4",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.7.0",
+    "@types/jsonwebtoken": "^9.0.9",
+    "@types/node": "22.15.29",
+    "@types/react": "19.1.6",
+    "@types/react-dom": "19.1.5",
+    "autoprefixer": "^10.4.21",
+    "dotenv": "^16.5.0",
+    "eslint": "^9.28.0",
+    "eslint-config-next": "^15.3.3",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-import-resolver-typescript": "^4.4.2",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-react": "^7.37.4",
-    "eslint-plugin-tailwindcss": "^3.17.5",
+    "eslint-plugin-prettier": "^5.4.1",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-tailwindcss": "^3.18.0",
     "esno": "^4.8.0",
-    "postcss": "^8.5.1",
+    "postcss": "^8.5.4",
     "postcss-html": "^1.8.0",
     "postcss-import": "^16.1.0",
     "postcss-scss": "^4.0.9",
-    "prettier": "^3.4.2",
-    "pretty-quick": "^4.0.0",
-    "prisma": "^6.2.1",
-    "sass": "^1.83.4",
-    "stylelint": "^16.13.1",
+    "prettier": "^3.5.3",
+    "pretty-quick": "^4.2.2",
+    "prisma": "^6.8.2",
+    "sass": "^1.89.1",
+    "stylelint": "^16.20.0",
     "stylelint-config-html": "^1.1.0",
     "stylelint-config-prettier": "^9.0.5",
-    "stylelint-config-recommended": "^14.0.1",
-    "stylelint-config-standard": "^36.0.1",
-    "stylelint-order": "^6.0.4",
-    "tailwindcss": "^3.4.17",
-    "typescript": "^5.7.3"
+    "stylelint-config-recommended": "^16.0.0",
+    "stylelint-config-standard": "^38.0.0",
+    "stylelint-order": "^7.0.0",
+    "tailwindcss": "^4.1.8",
+    "typescript": "^5.8.3"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
